roundtime shouldn't be negative anymore, also updated to long
git-svn-id: http://svn.warlock.cc/svn/warlock/warlock2/trunk@1007 bb130017-240f-0410-9613-9c2c827820fb

+++ stormfront/cc.warlock.core.stormfront/src/main/cc/warlock/core/stormfront/internal/PromptTagHandler.java
@@ -13,7 +13,7 @@ import cc.warlock.core.stormfront.xml.StormFrontAttributeList;
  */
 public class PromptTagHandler extends DefaultTagHandler {
 	
-	protected int currentTime = 0;
+	protected long currentTime = 0;
 	protected RoundtimeTagHandler roundtimeHandler;
 	protected IStormFrontClient client;
 	protected String prompt;
@@ -34,7 +34,7 @@ public class PromptTagHandler extends DefaultTagHandler {
 		
 		if (attributes.getValue("time") != null)
 		{
-			currentTime = Integer.parseInt(attributes.getValue("time"));
+			currentTime = Long.parseLong(attributes.getValue("time"));
 			if (roundtimeHandler.isWaitingForPrompt()) {
 				roundtimeHandler.processFollowingPrompt(currentTime);
 			}
